feat: add ads validation schema for ad submissions
